(IDEAPAD) => Added Registration Functionality

+++ base/views.py
@@ -4,6 +4,7 @@ from django.shortcuts import render,redirect
 from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login , logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.db.models import Q
 from .models import Room, Topic
@@ -13,8 +14,12 @@ from .forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -31,8 +36,26 @@ def loginPage(request):
             messages.error(request,"username and password does not exisit !!!")
 
 
-    context = {}
+    context = {"page":page}
     return render(request,'base/login_register.html',context)
+
+def registerUser(request):
+      
+        form = UserCreationForm()
+
+        if request.method == 'POST':
+            form = UserCreationForm(request.POST)
+            if form.is_valid():
+                user = form.save(commit=False)
+                user.username = user.username.lower()
+                user.save()
+                login(request,user)
+                return redirect('home')
+            else:
+                messages.error(request,'An error occured during Registrations')
+
+        context={"form":form}
+        return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
@@ -74,6 +97,9 @@ def createRoom(request):
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse('Your are not Allowed to Update This Room...')
     
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -86,6 +112,10 @@ def updateRoom(request, pk):
 @login_required(login_url="/login")
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    
+    if request.user != room.host :
+        return HttpResponse('Your are not Allowed to Delete This Room...')
     
     if request.method == 'POST':
         room.delete()
